refactor: Remove PageSearchableListLayout component
Removed the PageSearchableListLayout.vue component and PartnerAll.vue is reverted to original file.
