+++ src/partners/components/PartnerAll.vue
@@ -3,37 +3,57 @@ import { ref } from 'vue'
 import PartnerHero from './PartnerHero.vue'
 import PartnerList from './PartnerList.vue'
 import PartnerJoin from './PartnerJoin.vue'
-import data from '../partners.json'
 import { Partner } from './type'
-import PageSearchableListLayout from '@theme/components/PageSearchableListLayout.vue'
+import { VTIconSearch } from '@vue/theme'
 
+const query = ref('')
 
-const partners = ref<Partner[]>(data as Partner[])
+function filter(p: Partner): boolean {
+  return (
+    includes(p.name, query.value) ||
+    p.region.some((r) => includes(r, query.value))
+  )
+}
 
-const filterPartner = (partner: Partner, query: string): boolean => {
-  const lowerQuery = query.toLowerCase()
-  return partner.name.toLowerCase().includes(lowerQuery) ||
-    partner.region.some(region => region.toLowerCase().includes(lowerQuery))
+function includes(a: string, b: string) {
+  return a.toLowerCase().includes(b.toLowerCase())
 }
 </script>
 
 <template>
-  <PageSearchableListLayout
-    :title="'Browse All Partners'"
-    :items="partners"
-    searchPlaceholder="Search partners by name or region"
-    :filterFunction="filterPartner"
-  >
-    <template #hero="{ title }">
-      <PartnerHero :title="title" />
-    </template>
-
-    <template #list="{ filter }">
-      <PartnerList :filter="filter" />
-    </template>
-
-    <template #join>
-      <PartnerJoin />
-    </template>
-  </PageSearchableListLayout>
+  <PartnerHero title="Browser All Partners" />
+  <div class="container">
+    <VTIconSearch class="icon" />
+    <input
+      placeholder="Search partners by name or region"
+      v-model="query"
+    />
+    <PartnerList :filter="filter" />
+  </div>
+  <PartnerJoin />
 </template>
+
+<style scoped>
+input {
+  width: 100%;
+  padding: 8px 12px 8px 32px;
+  border-bottom: 1px solid var(--vt-c-divider-light);
+  margin-bottom: 2em;
+}
+
+.container {
+  max-width: 960px;
+  margin: 1em auto 2em;
+  padding: 0 28px;
+  position: relative;
+}
+
+.icon {
+  width: 18px;
+  height: 18px;
+  fill: var(--vt-c-text-3);
+  position: absolute;
+  top: 12px;
+  left: 32px;
+}
+</style>
